requirements.txt
Django>=1.7
django-countries>=2.1.2
django-durationfield>=0.5
django-extra-views>=0.6.5
django-localflavor>=1.0
djangorestframework>=3.3.1
# We need Python Markdown for django.contrib.markup. markdown2 is not enough.
Markdown>=2.4
python-dateutil>=2.2
stripe>=1.25.0
